# website/public/openapi.yaml
...
openapi: 3.1.0
servers:
- url: "{node}"
  description: "Any node of a deployment. Every node answers every endpoint. A node\
    \ that does not hold the index writer forwards the writes it receives, and answers\
    \ a search from the copy it holds."
  variables:
    node:
      default: http://localhost:8080
      description: "The origin a node is reached at, followed by the path prefix it\
        \ is served under if a proxy in front of it adds one. The tutorials and the\
        \ published images use `http://localhost:8080`. A deployment replaces it with\
        \ its own address."
components:
  schemas:
...
    AuditedGeneration:
      type: object
      description: One generation as the registry and storage each describe it.
      properties:
        name:
          type: string
          description: The name of the generation.
          examples:
          - "1"
        registered:
          type: boolean
          description: A boolean indicating whether the registry names the generation.
        stored:
          description: "What storage holds under it. `SYNCED`: storage holds a manifest;\
            \ nodes can pull and serve this generation. `INCOMPLETE`: storage holds\
            \ a prefix without a manifest (such as an unfinished push or what an interrupted\
            \ removal left of a deleted generation). `MISSING`: the generation is\
            \ registered, but nothing exists in storage."
          type: string
          $ref: "#/components/schemas/Stored"
        removedAt:
          type: string
          description: "When the generation was deleted on its own, as an ISO 8601\
            \ timestamp. Present while its storage waits for the sweep that removes\
            \ it. A generation of a deleted index carries the index's `removedAt`\
            \ instead. Omitted otherwise."
          examples:
          - 2026-09-03T10:15:00Z
    AuditedIndex:
      type: object
      description: One index as the registry and storage each describe it.
      properties:
        name:
          type: string
          description: Name of the index.
          examples:
          - products
        registered:
          type: boolean
          description: A boolean indicating whether the registry has an entry for
            the index.
        live:
          type: string
          description: The generation the index answers for. Omitted when unregistered
            or when no generation is live.
          examples:
          - "2"
        proposedLive:
          type: string
          description: "The generation that a repair with `promoteNewest` would make\
            \ live. Omitted when none would be promoted, which includes a deleted\
            \ index."
          examples:
          - "1"
        removedAt:
          type: string
          description: "When the index was deleted, as an ISO 8601 timestamp. Present\
            \ while the storage of the deleted index waits for the sweep that removes\
            \ it; a repair registers such an index only when asked to restore it.\
            \ Omitted otherwise."
          examples:
          - 2026-09-03T10:15:00Z
        generations:
          type: array
          items:
            $ref: "#/components/schemas/AuditedGeneration"
          description: "A list of generations found for the index, ordered by name."
...
    ErrorDetail:
      type: object
      description: One problem found in a request.
      properties:
        code:
          type: string
          description: The error code identifying this specific problem.
          examples:
          - index:field:invalid_primary_key_multiple
        message:
          type: string
          description: Human-readable description of this problem.
          examples:
          - "Field `id` is marked as a primary key and multiple, primary keys can\
            \ not have multiple values"
        path:
          type: string
          description: "Location of the offending value in the request, such as `fields.title`\
            \ or `documents[1].nonexistent`. Omitted when the problem applies to the\
            \ request as a whole."
          examples:
          - id
        arguments:
          type: object
          additionalProperties:
            type: string
          description: "The values the message was rendered with, so a client can\
            \ render a message of its own from the code."
    ErrorResponse:
      description: "The body of every failed request. Error codes use colon-separated\
        \ namespaces such as `index:field:invalid_name`, are stable across API versions,\
        \ and are never renamed or reused, so clients match on `code` rather than\
        \ on `message`. See [Errors](https://exofind.dev/reference/errors/)."
      type: object
      properties:
        code:
          type: string
          description: "Identifies the failure type. For validation failures, this\
            \ is `validation` and the individual problems are listed in `errors`.\
            \ For all other failures, this matches the `code` of the single entry\
            \ in `errors`."
          examples:
          - validation
        message:
          type: string
          description: "Human-readable message for log output. Match on `code` rather\
            \ than on `message`. When a validation failure contains one problem, this\
            \ is that problem's message. When it contains several, it reads `Request\
            \ contains N errors`."
          examples:
          - Request contains 2 errors
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
          description: "All problems found in the request. A validation failure reports\
            \ every field with a problem, so a caller can fix them in one pass rather\
            \ than one request at a time."
...
    Registry:
      type: string
      enum:
      - PRESENT
      - ABSENT
      - CORRUPT
    RegistryAuditResponse:
      description: "The registry compared with what remote storage holds. Reported\
        \ by the engine and never accepted as input. See [Audit](https://exofind.dev/reference/admin-api/#audit)."
      examples:
      - registry: PRESENT
        indexes:
        - name: products
          registered: true
          live: "2"
          generations:
          - name: "1"
            registered: true
            stored: SYNCED
          - name: "2"
            registered: true
            stored: SYNCED
        - name: staging
          registered: false
          removedAt: 2026-09-03T10:15:00Z
          generations:
          - name: "1"
            registered: false
            stored: SYNCED
        unusable: []
      type: object
      properties:
        registry:
          description: "The state of the registry object: `PRESENT`, `ABSENT` (no\
            \ registry object), or `CORRUPT` (contents cannot be parsed)."
          type: string
          $ref: "#/components/schemas/Registry"
        indexes:
          type: array
          items:
            $ref: "#/components/schemas/AuditedIndex"
          description: "Every index named by the registry or found in storage, ordered\
            \ by name."
        unusable:
          type: array
          items:
            type: string
          description: Storage prefixes whose names no index or generation may carry
            (as `index` or `index/generation`). A repair never registers these prefixes.
    RegistryRepairRequest:
      description: "How a repair should treat the indexes it creates, and which deleted\
        \ indexes or generations it should bring back. The entire body is optional."
      examples:
      - promoteNewest: true
        restore:
        - books
      type: object
      properties:
        promoteNewest:
          type: boolean
          description: "When `true`, each index created by the repair answers for\
            \ its highest-numbered generation. Hand-named generations are not selected.\
            \ Indexes that are already registered keep what they answer for. When\
            \ `false`, a created index answers for nothing until a generation is promoted."
          default: false
        restore:
          type: array
          items:
            type: string
          description: "Names of deleted indexes (`books`) or generations (`books@2`)\
            \ whose storage the sweep has not removed yet, to bring back. The repair\
            \ takes the removal mark off each one and registers what it holds like\
            \ any other unregistered storage. A name without a mark changes nothing.\
            \ Deleted storage is never registered without being named here."
          examples:
          - - books
    RegistryRepairResponse:
      description: "Summary of the changes made by a repair. Every list empty means\
        \ the registry already named everything storage holds, nothing was restored,\
        \ and nothing was written."
      examples:
      - createdIndexes:
        - products
        addedGenerations:
        - products@2
        promoted:
        - products@2
        restored:
        - products
      type: object
      properties:
        createdIndexes:
          type: array
          items:
            type: string
          description: "Index entries added to the registry, ordered by name."
        addedGenerations:
          type: array
          items:
            type: string
          description: "Generations added to the registry, formatted as `index@generation`\
            \ and ordered by name."
        promoted:
          type: array
          items:
            type: string
          description: "Generations made live by the repair, formatted as `index@generation`\
            \ and ordered by name."
        restored:
          type: array
          items:
            type: string
          description: "Deleted indexes and generations whose removal mark the repair\
            \ took off, as `index` or `index@generation`, in the order the request\
            \ named them. What they hold appears in the other lists."
...
    Stored:
      type: string
      enum:
      - SYNCED
      - INCOMPLETE
      - MISSING
...
  securitySchemes:
    apiKey:
      type: http
      description: "An API key sent as a bearer token, such as `Authorization: Bearer\
        \ exok_4ff6b760264c1918_ePQcdT1O9HSATZoXfDbT8hhHGsP9VpZH`. A key carries grants\
        \ that pair permissions with index patterns; the permission each endpoint\
        \ needs is named in its description. Nodes running with `EXOFIND_AUTH_MODE=none`\
        \ accept requests without a credential, and a node with `EXOFIND_AUTH_ANONYMOUS_KEY`\
        \ set serves requests that carry none with the permissions of that key."
      scheme: bearer
...
paths:
...
  /v1alpha1/admin/registry/actions/repair:
    post:
      summary: Repair the registry from storage
      description: |-
        Registers every `SYNCED` generation that storage holds and the registry does not name. The repair operation only adds entries: it keeps existing entries as stored and never deletes an index, a generation, or storage data. If the registry is absent, the repair writes it fresh. If the registry is corrupt, the repair replaces it with one rebuilt from storage.

        When `"promoteNewest": true`, each index created by the repair answers for its highest-numbered generation. Hand-named generations are not selected. Indexes that are already registered keep what they answer for. When omitted or false, created indexes answer for nothing until a generation is promoted.

        Storage of a deleted index or generation, which the audit reports with `removedAt`, is skipped until the sweep removes it. Naming it in `"restore"` takes the removal mark off and registers it like any other unregistered storage, which is how a delete is taken back within the grace period.

        The write is conditional and rebuilds on top of concurrent registry changes. The answering node applies the repaired registry immediately; other nodes pick it up within `EXOFIND_INDEXES_REFRESH_INTERVAL`.

        Served by whichever node receives the request and never forwarded. Answers only in object storage mode. Requires the `registry.repair` permission (deployment-scoped).
      operationId: repairRegistry
      tags:
      - Registry
      requestBody:
        content:
          application/json:
            examples:
              repair:
                value:
                  promoteNewest: true
                  restore:
                  - books
            schema:
              $ref: "#/components/schemas/RegistryRepairRequest"
        required: true
      responses:
        "200":
          description: A summary of the changes made by the repair.
          content:
            application/json:
              examples:
                repaired:
                  value:
                    createdIndexes:
                    - products
                    addedGenerations:
                    - products@2
                    promoted:
                    - products@2
                    restored:
                    - products
              schema:
                $ref: "#/components/schemas/RegistryRepairResponse"
        "401":
          description: The request carries no credential accepted by this node.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The credential does not have the `registry.repair` permission.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: "The endpoint was called on a node configured with local storage\
            \ (`index:registry:audit_unavailable`), or writing the repaired registry\
            \ failed (`index:registry:conflict`, `index:registry:io_error`). The registry\
            \ remains unchanged."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
      security:
      - apiKey: []
  /v1alpha1/admin/registry/audit:
    get:
      summary: Audit the registry against storage
      description: |-
        Reads the registry and remote storage, comparing the two without changing either.

        Unregistered generations in storage indicate interrupted rollouts, or deleted indexes and generations whose storage waits for the sweep that removes it; those carry `removedAt`. Registered generations missing from storage have no data available to pull.

        Served by whichever node receives the request and never forwarded. Answers only in object storage mode. Requires the `registry.audit` permission (deployment-scoped).
      operationId: auditRegistry
      tags:
      - Registry
      responses:
        "200":
          description: Comparison of the shared registry with remote storage.
          content:
            application/json:
              examples:
                audit:
                  value:
                    registry: PRESENT
                    indexes:
                    - name: products
                      registered: true
                      live: "2"
                      generations:
                      - name: "1"
                        registered: true
                        stored: SYNCED
                      - name: "2"
                        registered: true
                        stored: SYNCED
                    - name: staging
                      registered: false
                      removedAt: 2026-09-03T10:15:00Z
                      generations:
                      - name: "1"
                        registered: false
                        stored: SYNCED
                    unusable: []
              schema:
                $ref: "#/components/schemas/RegistryAuditResponse"
        "401":
          description: The request carries no credential accepted by this node.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The credential does not have the `registry.audit` permission.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The endpoint was called on a node configured with local storage
            rather than shared storage (`index:registry:audit_unavailable`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - apiKey: []
...
info:
  title: Exofind
  version: v1alpha1
  description: Search indexes kept in S3-compatible object storage. Every endpoint
    accepts and returns JSON.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0